moved handleParse to service
