+++ src/sections/staff/staffService.ts
@@ -6,6 +6,8 @@
  */
 
 import { supabaseClient } from "../../services/supabaseClient";
+import { v4 as uuid } from 'uuid';
+import { read, utils } from "xlsx";
 
 type Staff = {
     id: string,
@@ -30,6 +32,26 @@ class StaffService {
         if (error) {
             throw new Error(error.message);
         }
+    }
+
+    // referenced example at https://docs.sheetjs.com/docs/demos/frontend/react/
+    async handleParse(file: File) {
+        const arrayBuffer = await file.arrayBuffer();
+
+        const workbook = read(arrayBuffer);
+
+        const worksheet = workbook.Sheets[workbook.SheetNames[0]];
+        const data: Staff[] = utils.sheet_to_json(worksheet);
+
+        // modify excel sheet data
+        data.map((employee) => {
+            // add id and date fields
+            employee.id = uuid();
+            employee.created_at = new Date();
+            // convert roles string into an array
+            employee.roles = employee.roles.toString().split(",");
+        })
+        return data;
     }
 }
